Add workspace file
Create a multi-root workspace to begin with,
so that users can open this directly without
the surprising restart when adding their first DB.
Instead, the restart happens (more deliberately)
when they choose to open the workspace file.

Codespaces does not yet have a way to specify
which workspace file to open, so we remain reliant
on the user opening the folder and then being
prompted to open this workspace instead.