Add initial CodeTour
Ensure the CodeTour extension is installed.

Add a tour that walks the user through the basic steps of
running a query: show the DB, show the parts of the query,
and automatically run the query.

Automatic DB selection doesn't work yet,
because we need to find a way to pass the DB file URI
to the command.