Make sure we install packs compatible with our CLI
We've introduced some changes to run `codeql pack install` for both
the `tutorial-queries` QL pack and when you add a new database.

Unfortunately there is currently a danger that the packs we install
might not be compatible with the CLI version that you're running
in the codespace.

In the future we also want to port this behaviour of installing
dependencies for QL packs into the QL extension. Which means it will
have to be compatible with your local CLI version.

This manifest allows us to tell the `codeql pack install` command
to try to resolve packages from within the workspace.

Docs: https://docs.github.com/en/code-security/codeql-cli/codeql-cli-reference/about-codeql-workspaces#the-codeql-workspaceyml-file