Open extension on first step of code tour
Follow up to https://github.com/github/vscode-codeql/pull/2188.

We changed the extension to automatically open the code tour workspace
file when you click on it.

In the code tour, we open the extension on the sixth step. This is quite
late in the process, we want to have everything ready for the user from
step 1.

We want to add an extra step at the start of the tour where we open the
extension to force the workspace to be opened. This has an added
advantage that it will also display the pop-up to ask for permission to
record telemetry. So we get these issues out of the way on the first
step.

Side-note: if the user ignores the tour and goes off and clicks on the
extension on their own, we'll load the workspace for them (🎉 🎉 🎉),
so we're not tied to the tour to make the codespace behave as expected.