Detect newly created skeleton query folders
We are creating new folders when you generate queries via the
"CodeQL: Create Query" command.

These will be created in a QL pack folder called
`codeql-custom-queries-<language>`.

In https://github.com/github/codespaces-codeql/pull/26 we made it possible
to detect when we install dependencies for the tutorial-queries QL pack.

We'd like to do the same for our generated QL packs.